feat(F89): F89↔F87 bridge claim - AT-lock inherits F87-Truly classification
New typed bridge claim F89F87TrulyInheritance under
compute/RCPsiSquared.Diagnostics/F87/. Tier-1-Derived structural identity:
F89's per-bond Hamiltonian H_b = J·(X_pX_q + Y_pY_q) is exactly the
canonical F87 "XX+YY" witness, which classifies as TrichotomyClass.Truly.
F87-Truly is the bit-exact operator equation Π·L·Π⁻¹ + L + 2σ·I = 0,
which forces the Absorption Theorem to hold exactly: Re(λ) = −2γ₀·⟨n_XY⟩
for every Liouvillian eigenmode. F89's F_a modes live in the overlap
subspace where n_diff = 1, giving the F89 AT-lock Re(λ_n) = −2γ₀ exactly
(as typed in F89PathKAtLockMechanismClaim).

Methods exposed:
- F89BondTerms: static [XX, YY] bond-term specification
- ClassifyF89Hamiltonian(chain) → TrichotomyClass via PauliPairTrichotomy
- IsAtLockPreconditionSatisfied(chain) → bool (= classify == Truly)

Documents the bond-class extension: Heisenberg XX+YY+ZZ also F87-Truly
under Z-deph (ZZ is Π²-even-truly), preserves AT-lock. Documents the
break case: replacing any bond's XX+YY with YZ+ZY (F87-Soft, EQ-030
Marrakesh anchor) mixes the spectrum and breaks the F89 AT-lock.

Tests: 8/8 grün at N=3..5 (full-L Evd cheap at small N; classification
is N-invariant for Pauli-pair bond Hamiltonians).

Drive-by: long→int cast in F86VerificationExploration.cs lines 125, 128
(StatesP/StatesQ recently changed to long[]).

Closes F89↔F87 bridge gap identified in 2026-05-16 inventory.

+++ compute/RCPsiSquared.Diagnostics.Tests/Foundation/F86VerificationExploration.cs
@@ -122,10 +122,10 @@ public class F86VerificationExploration
         var A = ComplexMatrix.Build.Dense(d, d);
         for (int i = 0; i < block.Basis.Mp; i++)
         {
-            int p = block.Basis.StatesP[i];
+            int p = (int)block.Basis.StatesP[i];
             for (int j = 0; j < block.Basis.Mq; j++)
             {
-                int q = block.Basis.StatesQ[j];
+                int q = (int)block.Basis.StatesQ[j];
                 Complex coeff = v[i * block.Basis.Mq + j];
                 A[p, q] += coeff;
                 A[q, p] += Complex.Conjugate(coeff);
